+++ Planner/EnergyPlanner.cpp
@@ -4,7 +4,7 @@
 
 #include "EnergyPlanner.h"
 
-EnergyPlanner::EnergyPlanner(int numPoints, float epsilon) : Planner(numPoints, epsilon) {}
+EnergyPlanner::EnergyPlanner(int numPoints, double epsilon) : Planner(numPoints, epsilon) {}
 
 double EnergyPlanner::cost(RobotState *startState, RobotState *endState) {
     // todo: fill this in!
