+++ Planner/Planner.h
@@ -12,7 +12,7 @@
 
 class Planner {
 public:
-    Planner(int numPoints, float epsilon);
+    Planner(int numPoints, double epsilon);
     void findBestPath();
     virtual ~Planner(); // todo: smart pointers? (avoid deletion)
 
@@ -36,7 +36,7 @@ private:
 
     RobotState* root;
     int maxIterations;
-    float epsilon;
+    double epsilon;
     Location start;
     Location end;
     ShapeDrawer drawer;
